+++ nx-loki/src/utils/versions.ts
@@ -1,0 +1,1 @@
+export const LOKI_VERSION = "0.35.1";