+++ CHANGELOG.md
@@ -1,3 +1,89 @@
+## 0.0.9 (2025-03-23)
+
+### 🚀 Features
+
+- remove prompt in schema ([a3a41f0](https://github.com/muhammedgaygisiz/nx-loki/commit/a3a41f0))
+- add loki as dev dependency ([6d8c8ce](https://github.com/muhammedgaygisiz/nx-loki/commit/6d8c8ce))
+
+### ❤️ Thank You
+
+- e529454
+
+## 0.0.8 (2025-03-21)
+
+### 🩹 Fixes
+
+- fix url in generators.json ([a4be6c1](https://github.com/muhammedgaygisiz/nx-loki/commit/a4be6c1))
+
+### ❤️ Thank You
+
+- e529454
+
+## 0.0.7 (2025-03-21)
+
+### 🩹 Fixes
+
+- fix location of generators entry in package.json ([9428f4b](https://github.com/muhammedgaygisiz/nx-loki/commit/9428f4b))
+
+### ❤️ Thank You
+
+- e529454
+
+## 0.0.6 (2025-03-21)
+
+This was a version bump only, there were no code changes.
+
+## <small>0.0.7 (2025-03-21)</small>
+
+## <small>0.0.6 (2025-03-20)</small>
+
+- chore: bump to version 0.0.4 :rocket: ([401e9ce](https://github.com/muhammedgaygisiz/nx-loki/commit/401e9ce))
+- chore: bump to version 0.0.6 :rocket: ([b7f1583](https://github.com/muhammedgaygisiz/nx-loki/commit/b7f1583))
+- chore: fix path to generators ([df43d0e](https://github.com/muhammedgaygisiz/nx-loki/commit/df43d0e))
+
+## <small>0.0.6 (2025-03-20)</small>
+
+- chore: bump to version 0.0.4 :rocket: ([401e9ce](https://github.com/muhammedgaygisiz/nx-loki/commit/401e9ce))
+- chore: fix path to generators ([df43d0e](https://github.com/muhammedgaygisiz/nx-loki/commit/df43d0e))
+
+## <small>0.0.5 (2025-03-20)</small>
+
+- chore(release): publish 0.0.5 ([50c481f](https://github.com/muhammedgaygisiz/nx-loki/commit/50c481f))
+
+## <small>0.0.4 (2025-03-20)</small>
+
+- chore: fix path to generators ([df43d0e](https://github.com/muhammedgaygisiz/nx-loki/commit/df43d0e))
+
+## <small>0.0.5 (2025-03-20)</small>
+
+- chore(release): publish 0.0.5 ([50c481f](https://github.com/muhammedgaygisiz/nx-loki/commit/50c481f))
+
+## 0.0.5 (2025-03-20)
+
+This was a version bump only, there were no code changes.
+
+## <small>0.0.3 (2025-03-20)</small>
+
+- chore: move files to src ([1e8579f](https://github.com/muhammedgaygisiz/nx-loki/commit/1e8579f))
+
+## <small>0.0.4 (2025-03-20)</small>
+
+- chore(release): publish 0.0.4 ([399b811](https://github.com/muhammedgaygisiz/nx-loki/commit/399b811))
+
+## 0.0.4 (2025-03-20)
+
+This was a version bump only, there were no code changes.
+
+## 0.0.3 (2025-03-20)
+
+### 🚀 Features
+
+- initial impl of init generator ([7a91cd9](https://github.com/muhammedgaygisiz/nx-loki/commit/7a91cd9))
+
+### ❤️ Thank You
+
+- e529454
+
 ## <small>0.0.2 (2025-03-18)</small>
 
 - chore: setup tag message ([e035e36](https://github.com/muhammedgaygisiz/nx-loki/commit/e035e36))